Update TADS to 3.1.3 / 2.5.17.
This incorporates FrobTADS code to handle mostt OS-specific operations.

See #203.

+++ tads/tads3/test/test_utf8.cpp
@@ -54,7 +54,8 @@ int main()
 
     /* compute the length */
     len = p.lenz();
-    printf("test1: len = %d - %s\n", len,
+    printf("test1: len = %lu - %s\n",
+           (unsigned long)len,
            (len == sizeof(test1)/sizeof(test1[0]) - 1) ? "ok" : "ERROR");
 
     /* 
@@ -68,8 +69,8 @@ int main()
         ch = p.getch();
 
         /* check it */
-        printf("ch[%d] = %04x - %s\n", i, ch,
-               (ch == *wp ? "ok" : "ERROR"));
+        printf("ch[%lu] = %04x - %s\n",
+               (unsigned long)i, ch, (ch == *wp ? "ok" : "ERROR"));
     }
 
     /* go backwards through the buffer */
@@ -85,8 +86,8 @@ int main()
         ch = p.getch();
 
         /* check it */
-        printf("ch[%d] = %04x - %s\n", i, ch,
-               (ch == *wp ? "ok" : "ERROR"));
+        printf("ch[%lu] = %04x - %s\n",
+               (unsigned long)i, ch, (ch == *wp ? "ok" : "ERROR"));
     } while (i != 0);
 
     /* success */
